Import A346BXXUBEYI7 OSRC kernel source

+++ kernel/kernel_device_modules-6.6/drivers/gpu/drm/mediatek/mml/mtk-mml-adaptor.c
@@ -575,6 +575,10 @@ int mml_ctx_init(struct mml_ctx *ctx, struct mml_dev *mml,
 
 	}
 	ctx->wq_destroy = alloc_ordered_workqueue("%s", 0, threads[1]);
+	if (!ctx->wq_destroy) {
+		mml_err("[adpt]fail to create destroy workqueue %s", threads[1]);
+		goto err;
+	}
 	if (threads[2]) {
 		ctx->kt_config[0] = kthread_create_worker(0, "%s", threads[2]);
 		if (IS_ERR(ctx->kt_config[0])) {
@@ -638,8 +642,11 @@ void mml_ctx_deinit(struct mml_ctx *ctx)
 		frame_config_queue_destroy(cfg);
 	}
 
-	destroy_workqueue(ctx->wq_destroy);
-	ctx->wq_destroy = NULL;
+	mml_msg("[adpt]%s destroy_workqueue %p on ctx %p", __func__, ctx->wq_destroy, ctx);
+	if (ctx->wq_destroy) {
+		destroy_workqueue(ctx->wq_destroy);
+		ctx->wq_destroy = NULL;
+	}
 	if (ctx->kt_config[0]) {
 		kthread_destroy_worker(ctx->kt_config[0]);
 		ctx->kt_config[0] = NULL;
@@ -648,8 +655,10 @@ void mml_ctx_deinit(struct mml_ctx *ctx)
 		kthread_destroy_worker(ctx->kt_config[1]);
 		ctx->kt_config[1] = NULL;
 	}
-	kthread_destroy_worker(ctx->kt_done);
-	ctx->kt_done = NULL;
+	if (ctx->kt_done) {
+		kthread_destroy_worker(ctx->kt_done);
+		ctx->kt_done = NULL;
+	}
 
 	for (i = 0; i < ARRAY_SIZE(ctx->tile_cache); i++)
 		for (j = 0; j < ARRAY_SIZE(ctx->tile_cache[i].func_list); j++)
